pyproject.toml
[project]
name = "dog-breed-classifier"
version = "1.0.0"
description = "Dog Breed Classification & Human-to-Dog Similarity Experiment"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "Pillow>=10.0.0",
    "kaggle>=1.5.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-timeout>=2.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
